cloud-functions/image-upload/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.14
google-cloud-firestore>=2.16
